make index.js revampIndex, create folder for clients
